Find players by season rather than matches played
This is the obvious way to do it. Before I was just looking for all the
players that played a match in the season before listing them. Now in
all views the players are queried by the season they belong to,
including the Season rankings view, the opponents drop-down list
when reporting games, etc.

Also snuck in an easier way to turn on admin view.

And fixed season view to sort by First Name, Last Name for ties.
This makes sense since the names are printed as first name then
last name and matches the drop-down opponents list in reporting view.